% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Obs_v_Fitted.R
\name{Plot_Obs_v_Fitted}
\alias{Plot_Obs_v_Fitted}
\title{FUNCTION_TITLE}
\usage{
Plot_Obs_v_Fitted(df, Var, weight, unitpred, response, LUTable = NA,
  LUType = "Direct", LUOrdered = "N")
}
\arguments{
\item{df}{Dataframe}

\item{Var}{Variable}

\item{weight}{Weight}

\item{unitpred}{Per Unit Weight Prediction}

\item{response}{Response field}

\item{LUTable}{Lookup table, Default: NA}

\item{LUType}{Lookup type, Default: 'Direct'}

\item{LUOrdered}{IS LU Ordered?, Default: 'N'}
}
\value{
Plot in plot
}
\description{
FUNCTION_DESCRIPTION
This is the top bit of text
}
\details{
Some details here
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
